Changed configs of costmpas for mapupdate
Changed the used sensors and LiDAR-Topics for better results of the updated maps (now we don't use the merged LiDAR-Scan "scan_wcr_averaged" but we use the front- and back-scan to update the generated costmaps). This helps to keep false clearing of objects (especially the clearing of walls in the unoccupied costmaps) low. Also we lowered the update rate of the costmaps. This helps to use the mapudpate on the raspberry pi.